PP-5605: refactor HistoricalEventEmitterWorker
extract emitting from Worker to HistoricalEventEmitter
